@@ -2,8 +2,13 @@
   "name": "phpro/zf-doctrine-hydration-module",
   "license": "MIT",
   "description": "Doctrine hydrators for ZF2 and ZF3",
+  "scripts": {
+    "test": "phpunit",
+    "cs": "phpcs",
+    "psalm": "psalm --long-progress --show-info=false --output-format=github"
+  },
   "keywords": [
-    "zf",
+    "laminas",
     "doctrine",
     "hydrator"
   ],
@@ -19,21 +24,22 @@
     }
   },
   "require": {
-    "php": "^5.6 || ^7.0",
-    "doctrine/common": "^2.6.1",
-    "doctrine/doctrine-module": "^1.2 || ^2.1.7",
+    "php": "^7.3 || ^8.0",
+    "doctrine/common": "^2.6.1 || ^3.0.0",
+    "doctrine/doctrine-module": "^3.0.1 || ^4.0 || ^4.1",
     "doctrine/instantiator": "^1.0.5",
-    "zendframework/zend-servicemanager": "^3.3.2",
-    "zendframework/zend-hydrator": "^1.1 || ^2.2.1",
-    "zendframework/zend-modulemanager": "^2.7.2"
+    "laminas/laminas-servicemanager": "^3.3.2",
+    "laminas/laminas-hydrator": "^2.2.1 || ^3.2.0 || ^4.0.0",
+    "laminas/laminas-modulemanager": "^2.7.2"
   },
   "require-dev": {
-    "fabpot/PHP-CS-Fixer": "^1.11.6",
-    "phpunit/phpunit": "^5.7.27",
-    "doctrine/doctrine-orm-module": "^1.1 || ^2.1.2",
-    "doctrine/doctrine-mongo-odm-module": "^0.11 || ^1.0",
-    "doctrine/mongodb-odm": "^1.1",
-    "phpro/grumphp": "^0.9.1"
+    "phpunit/phpunit": "^9.3",
+    "doctrine/doctrine-orm-module": "^3.0",
+    "doctrine/doctrine-mongo-odm-module": "^2.0 || ^3.0",
+    "doctrine/mongodb-odm": "^2.1",
+    "vimeo/psalm": "^4.1",
+    "psalm/plugin-phpunit": "^0.13.0",
+    "squizlabs/php_codesniffer": "^3.5"
   },
   "autoload": {
     "psr-4": {
