@@ -1,7 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PhproTest\DoctrineHydrationModule\Tests\Hydrator;
 
+use Laminas\Hydrator\HydratorInterface;
 use Phpro\DoctrineHydrationModule\Hydrator\DoctrineHydrator;
 use PHPUnit\Framework\TestCase;
 
@@ -18,8 +21,8 @@ class DoctrineHydratorTest extends TestCase
      */
     protected function createHydrator($hydrateService = null, $extractService = null)
     {
-        $hydrateService = $hydrateService ? $hydrateService : $this->getMockBuilder('Zend\Hydrator\HydratorInterface')->getMock();
-        $extractService = $extractService ? $extractService : $this->getMockBuilder('Zend\Hydrator\HydratorInterface')->getMock();
+        $hydrateService = $hydrateService ? $hydrateService : $this->getMockBuilder('Laminas\Hydrator\HydratorInterface')->getMock();
+        $extractService = $extractService ? $extractService : $this->getMockBuilder('Laminas\Hydrator\HydratorInterface')->getMock();
 
         return new DoctrineHydrator($extractService, $hydrateService);
     }
@@ -39,7 +42,7 @@ class DoctrineHydratorTest extends TestCase
     public function it_should_have_a_hydrator_service()
     {
         $hydrator = $this->createHydrator();
-        $this->assertInstanceOf('Zend\Hydrator\HydratorInterface', $hydrator->getHydrateService());
+        $this->assertInstanceOf('Laminas\Hydrator\HydratorInterface', $hydrator->getHydrateService());
     }
 
     /**
@@ -48,7 +51,7 @@ class DoctrineHydratorTest extends TestCase
     public function it_should_have_an_extractor_service()
     {
         $hydrator = $this->createHydrator();
-        $this->assertInstanceOf('Zend\Hydrator\HydratorInterface', $hydrator->getExtractService());
+        $this->assertInstanceOf('Laminas\Hydrator\HydratorInterface', $hydrator->getExtractService());
     }
 
     /**
@@ -57,8 +60,8 @@ class DoctrineHydratorTest extends TestCase
     public function it_should_extract_an_object()
     {
         $object = new \stdClass();
-        $extracted = array('extracted' => true);
-        $extractService = $this->getMockBuilder('Zend\Hydrator\HydratorInterface')->getMock();
+        $extracted = ['extracted' => true];
+        $extractService = $this->getMockBuilder(HydratorInterface::class)->getMock();
         $extractService
             ->expects($this->any())
             ->method('extract')
@@ -76,9 +79,9 @@ class DoctrineHydratorTest extends TestCase
     public function it_should_hydrate_an_object()
     {
         $object = new \stdClass();
-        $data = array('field' => 'value');
+        $data = ['field' => 'value'];
 
-        $hydrateService = $this->getMockBuilder('Zend\Hydrator\HydratorInterface')->getMock();
+        $hydrateService = $this->getMockBuilder(HydratorInterface::class)->getMock();
         $hydrateService
             ->expects($this->any())
             ->method('hydrate')
@@ -97,14 +100,14 @@ class DoctrineHydratorTest extends TestCase
     public function it_should_use_a_generated_doctrine_hydrator_while_hydrating_an_object()
     {
         $object = new \stdClass();
-        $data = array('field' => 'value');
+        $data = ['field' => 'value'];
 
-        $hydrateService = $this->getMockBuilder('Doctrine\ODM\MongoDB\Hydrator\HydratorInterface')->getMock();
+        $hydrateService = $this->getMockBuilder(\Doctrine\ODM\MongoDB\Hydrator\HydratorInterface::class)->getMock();
         $hydrateService
             ->expects($this->any())
             ->method('hydrate')
             ->with($object, $data)
-            ->will($this->returnValue($object));
+            ->will($this->returnValue($data));
 
         $hydrator = $this->createHydrator($hydrateService, null);
         $result = $hydrator->hydrate($data, $object);
