@@ -5,7 +5,7 @@ namespace PhproTest\DoctrineHydrationModule\Tests\Hydrator\ODM\MongoDB\Strategy;
 use Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy\DefaultRelation as StrategyDefaultRelation;
 use PhproTest\DoctrineHydrationModule\Fixtures\ODM\MongoDb\HydrationEmbedMany;
 use PhproTest\DoctrineHydrationModule\Fixtures\ODM\MongoDb\HydrationUser;
-use Zend\Hydrator\Strategy\StrategyInterface;
+use Laminas\Hydrator\Strategy\StrategyInterface;
 
 /**
  * Class EmbeddedFieldTest.
@@ -32,7 +32,7 @@ class DefaultRelationTest extends AbstractMongoStrategyTest
         $embedded = new HydrationEmbedMany();
         $embedded->setId(1);
         $embedded->setName('name');
-        $user->addEmbedMany(array($embedded));
+        $user->addEmbedMany([$embedded]);
 
         $strategy = $this->getStrategy($this->dm, $user, 'embedMany');
         $result = $strategy->extract($user->getEmbedMany());
@@ -52,9 +52,9 @@ class DefaultRelationTest extends AbstractMongoStrategyTest
         $embedded->setId(1);
         $embedded->setName('name');
 
-        $data = array(
+        $data = [
             $embedded
-        );
+        ];
 
         $strategy = $this->getStrategy($this->dm, $user, 'embedMany');
         $strategy->hydrate($data);
