@@ -1,11 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy;
 
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ObjectManager;
 use DoctrineModule\Persistence\ObjectManagerAwareInterface;
-use DoctrineModule\Stdlib\Hydrator\Strategy\AbstractCollectionStrategy;
-use DoctrineModule\Stdlib\Hydrator\Strategy\AllowRemoveByValue;
+use Doctrine\Laminas\Hydrator\Strategy\AbstractCollectionStrategy;
+use Doctrine\Laminas\Hydrator\Strategy\AllowRemoveByValue;
 use Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\DoctrineObject;
 
 /**
@@ -23,7 +25,7 @@ abstract class AbstractMongoStrategy extends AbstractCollectionStrategy implemen
      *
      * @param ObjectManager $objectManager
      */
-    public function setObjectManager(ObjectManager $objectManager)
+    public function setObjectManager(ObjectManager $objectManager): void
     {
         $this->objectManager = $objectManager;
     }
@@ -33,15 +35,15 @@ abstract class AbstractMongoStrategy extends AbstractCollectionStrategy implemen
      *
      * @return ObjectManager
      */
-    public function getObjectManager()
+    public function getObjectManager(): ObjectManager
     {
         return $this->objectManager;
     }
 
     /**
-     * @param ObjectManager $objectManager Possibly injected by hydrator factory
+     * @param ObjectManager|null $objectManager Possibly injected by hydrator factory
      */
-    public function __construct($objectManager = null)
+    public function __construct(?ObjectManager $objectManager = null)
     {
         if ($objectManager) {
             $this->setObjectManager($objectManager);
@@ -70,7 +72,7 @@ abstract class AbstractMongoStrategy extends AbstractCollectionStrategy implemen
         $strategy->setClassMetadata($this->getClassMetadata());
         $strategy->setCollectionName($this->getCollectionName());
 
-        return $strategy->hydrate($value);
+        return $strategy->hydrate($value, null);
     }
 
     /**
