@@ -1,8 +1,8 @@
 <?php
 
-namespace Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy;
+declare(strict_types=1);
 
-use Doctrine\Common\Collections\Collection;
+namespace Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy;
 
 /**
  * Class PersistentCollection.
@@ -10,23 +10,17 @@ use Doctrine\Common\Collections\Collection;
 class DefaultRelation extends AbstractMongoStrategy
 {
     /**
-     * @param mixed $value
-     *
-     * @return array|mixed
-     *
-     * @throws \Exception
+     * {@inheritDoc}
      */
-    public function extract($value)
+    public function extract($value, ?object $object = null)
     {
         return $value;
     }
 
     /**
-     * @param mixed $value
-     *
-     * @return array|Collection|mixed
+     * @inheritDoc
      */
-    public function hydrate($value)
+    public function hydrate($value, $data)
     {
         // Beware of the collection strategies:
         $collection = $this->collectionName;
