@@ -1,8 +1,8 @@
 <?php
 
-namespace Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy;
+declare(strict_types=1);
 
-use Doctrine\Common\Collections\Collection;
+namespace Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy;
 
 /**
  * Class PersistentCollection.
@@ -10,11 +10,9 @@ use Doctrine\Common\Collections\Collection;
 class ReferencedField extends AbstractMongoStrategy
 {
     /**
-     * @param mixed $value
-     *
-     * @return mixed
+     * {@inheritDoc}
      */
-    public function extract($value)
+    public function extract($value, ?object $object = null)
     {
         if (!is_object($value)) {
             return $value;
@@ -22,7 +20,7 @@ class ReferencedField extends AbstractMongoStrategy
 
         $idField = $this->metadata->getIdentifier();
         $idField = is_array($idField) ? current($idField) : $idField;
-        $getter = 'get'.ucfirst($idField);
+        $getter = 'get' . ucfirst($idField);
 
         // Validate object:
         $rc = new \ReflectionClass($value);
@@ -34,11 +32,9 @@ class ReferencedField extends AbstractMongoStrategy
     }
 
     /**
-     * @param mixed $value
-     *
-     * @return array|Collection|mixed
+     * @inheritDoc
      */
-    public function hydrate($value)
+    public function hydrate($value, $data)
     {
         if (is_object($value)) {
             return $value;
