@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Phpro\DoctrineHydrationModule;
 
 class Module
@@ -9,6 +11,6 @@ class Module
      */
     public function getConfig()
     {
-        return include __DIR__.'/../config/module.config.php';
+        return include __DIR__ . '/../config/module.config.php';
     }
 }
