@@ -1,12 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PhproTest\DoctrineHydrationModule;
 
 error_reporting(E_ALL | E_STRICT);
-define('PROJECT_BASE_PATH', __DIR__.'/../..');
-define('TEST_BASE_PATH', __DIR__.'/..');
+define('PROJECT_BASE_PATH', __DIR__ . '/../..');
+define('TEST_BASE_PATH', __DIR__ . '/..');
 
-$autoloadFile = PROJECT_BASE_PATH.'/vendor/autoload.php';
+$autoloadFile = PROJECT_BASE_PATH . '/vendor/autoload.php';
 if (!file_exists($autoloadFile)) {
     throw new \RuntimeException('Install dependencies to run test suite.');
 }
@@ -43,12 +45,12 @@ class Bootstrap
      */
     protected function initAutoLoading()
     {
-        $this->autoLoader->addPsr4('PhproTest\\DoctrineHydrationModule\\Tests\\', __DIR__.'/Tests/');
-        $this->autoLoader->addPsr4('PhproTest\\DoctrineHydrationModule\\Fixtures\\', __DIR__.'/Fixtures/');
+        $this->autoLoader->addPsr4('PhproTest\\DoctrineHydrationModule\\Tests\\', __DIR__ . '/Tests/');
+        $this->autoLoader->addPsr4('PhproTest\\DoctrineHydrationModule\\Fixtures\\', __DIR__ . '/Fixtures/');
 
-        $this->autoLoader->addClassMap(array(
-            'Doctrine\\ODM\\MongoDB\\Tests\\BaseTest' => PROJECT_BASE_PATH.'/vendor/doctrine/mongodb-odm/tests/Doctrine/ODM/MongoDB/Tests/BaseTest.php',
-        ));
+        $this->autoLoader->addClassMap([
+            'Doctrine\\ODM\\MongoDB\\Tests\\BaseTest' => PROJECT_BASE_PATH . '/vendor/doctrine/mongodb-odm/tests/Doctrine/ODM/MongoDB/Tests/BaseTest.php',
+        ]);
     }
 
     /**
@@ -61,7 +63,7 @@ class Bootstrap
         define('DOCTRINE_MONGODB_SERVER', 'mongodb://localhost:27017');
 
         // Load annotated classes
-        \Doctrine\ODM\MongoDB\Mapping\Driver\AnnotationDriver::registerAnnotationClasses();
+        //\Doctrine\ODM\MongoDB\Mapping\Driver\AnnotationDriver::registerAnnotationClasses();
     }
 }
 
