@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PhproTest\DoctrineHydrationModule\Tests\Hydrator\ODM\MongoDB\Strategy;
 
 use Phpro\DoctrineHydrationModule\Hydrator\ODM\MongoDB\Strategy\DateTimeField;
@@ -35,7 +37,7 @@ class DateTimeFieldTest extends TestCase
     public function it_should_be_a_strategy_interface()
     {
         $strategy = $this->createStrategy();
-        $this->assertInstanceOf('Zend\Hydrator\Strategy\StrategyInterface', $strategy);
+        $this->assertInstanceOf('Laminas\Hydrator\Strategy\StrategyInterface', $strategy);
     }
 
     /**
